+++ backend/app/schemas/user_schema.py
@@ -9,3 +9,12 @@ class UserRegister(BaseModel):
 class UserLogin(BaseModel):
 	email: str
 	password: str
+
+
+class UserOut(BaseModel):
+	id: int
+	email: str
+	role: str
+
+	class Config:
+		from_attributes = True